Move rotation_classes.py to translation_classes.py
